// types/species.ts
export interface Species {
   average_height: string;
   average_lifespan: string;
   classification: string;
   created: Date;
   designation: string;
   edited: Date;
   eye_colors: string;
   films: string[];
   hair_colors: string;
   homeworld: string;
   language: string;
   name: string;
   people: string[];
   skin_colors: string;
   url: string;
}